Relative UI Menu + TextMeshPro
Made the menu UI relative in order to fit for all screen resolutions and ratios.

Replaced Text elements in the menu with TextMeshPro elements because of performance improvements and better layouting.

However, TextMeshPro has to be imported in the Assets. Therefore, I added these files to the .gitignore, because unity will ask you to import them anyway (the normal import without extras is sufficient) and they don't contain project specific information.